Information about 9 species of carnivores, derived from msleep

A dataset containing information about 9 species of carnivores. This dataset was derived from ggplot2::msleep.

carnivores

Format

A tibble (data frame) with 9 rows and 4 variables:

name

common name of the carnivore species

genus

the genus classification of the carnivore species

awake

the number of hours, on average, this species spends awake per day

brainwt

the species' average brain weight in kg, where several values are missing (NA)

Source

https://ggplot2.tidyverse.org/

Examples

# \dontrun{
  carnivores

#> # A tibble: 9 × 4
#>   name              genus        awake brainwt
#>   <chr>             <chr>        <dbl>   <dbl>
#> 1 Arctic fox        Vulpes        11.5  0.0445
#> 2 Cheetah           Acinonyx      11.9 NA     
#> 3 Dog               Canis         13.9  0.07  
#> 4 Gray seal         Haliochoerus  17.8  0.325 
#> 5 Jaguar            Panthera      13.6  0.157 
#> 6 Lion              Panthera      10.5 NA     
#> 7 Northern fur seal Callorhinus   15.3 NA     
#> 8 Red fox           Vulpes        14.2  0.0504
#> 9 Tiger             Panthera       8.2 NA     

# }