An updated and expanded version of the mammals sleep dataset, exported from ggplot2

This is the msleep dataset exported from the ggplot2 package version 3.3.5. For more information, see ?ggplot2::msleep.

msleep

Format

A tibble (data frame) with 83 rows and 11 variables:

name

common name

genus

the Genus classification of the species

vore

carnivore, omnivore, herbivore, or insectivore?

order

the Order classification of the species

conservation

the conservation status of the species

sleep_total

total amount of sleep, in hours

sleep_rem

rem sleep, in hours

sleep_cycle

length of sleep cycle, in hours

awake

amount of time spent awake, in hours

brainwt

brain weight in kilograms

bodywt

body weight in kilograms

Source

https://ggplot2.tidyverse.org/

Examples

# \dontrun{
  msleep
#> # A tibble: 83 × 11
#>    name   genus vore  order conservation sleep_total sleep_rem sleep_cycle awake
#>    <chr>  <chr> <chr> <chr> <chr>              <dbl>     <dbl>       <dbl> <dbl>
#>  1 Cheet… Acin… carni Carn… lc                  12.1      NA        NA      11.9
#>  2 Owl m… Aotus omni  Prim… NA                  17         1.8      NA       7  
#>  3 Mount… Aplo… herbi Rode… nt                  14.4       2.4      NA       9.6
#>  4 Great… Blar… omni  Sori… lc                  14.9       2.3       0.133   9.1
#>  5 Cow    Bos   herbi Arti… domesticated         4         0.7       0.667  20  
#>  6 Three… Brad… herbi Pilo… NA                  14.4       2.2       0.767   9.6
#>  7 North… Call… carni Carn… vu                   8.7       1.4       0.383  15.3
#>  8 Vespe… Calo… NA    Rode… NA                   7        NA        NA      17  
#>  9 Dog    Canis carni Carn… domesticated        10.1       2.9       0.333  13.9
#> 10 Roe d… Capr… herbi Arti… lc                   3        NA        NA      21  
#> # … with 73 more rows, and 2 more variables: brainwt <dbl>, bodywt <dbl>
# }